user context & game context
